requirements.txt
streamlit
tensorflow
tensorflow-hub
numpy
Pillow